feat: Player 클래스 생성
