+++ src/main/java/domain/Participant.java
@@ -65,43 +65,7 @@ public class Participant {
 		return bestScore;
 	}
 
-	public Versus initCompare(boolean isDealerBlackJack) {
-		boolean isParticipantBlackJack = isBlackJack();
-		if (isDealerBlackJack && isParticipantBlackJack) {
-			return Versus.DRAW;
-		}
-		if (!isDealerBlackJack && isParticipantBlackJack) {
-			return Versus.WIN;
-		}
-		return Versus.LOSE;
-	}
-
-	public Versus finalCompare(Participant other) {
-		if (isBust()) {
-			return Versus.LOSE;
-		}
-		if (other.isBust()) {
-			return Versus.WIN;
-		}
-		return whoIsHigh(other.getBestScore());
-	}
-
-	private Versus whoIsHigh(int otherScore) {
-		int myScore = getBestScore();
-		if (myScore > otherScore) {
-			return Versus.WIN;
-		}
-		if (myScore == otherScore) {
-			return Versus.DRAW;
-		}
-		return Versus.LOSE;
-	}
-
 	protected int getAceCount() {
 		return (int)hand.stream().filter(Card::isAce).count();
-	}
-
-	public boolean compareName(String name) {
-		return this.name.equals(name);
 	}
 }
