+++ src/main/java/domain/Player.java
@@ -1,0 +1,45 @@
+package domain;
+
+import java.util.List;
+
+public class Player extends Participant {
+	public Player(String name, List<Card> hand) {
+		super(name, hand);
+	}
+	
+	public Versus initCompare(boolean isDealerBlackJack) {
+		boolean isParticipantBlackJack = isBlackJack();
+		if (isDealerBlackJack && isParticipantBlackJack) {
+			return Versus.DRAW;
+		}
+		if (!isDealerBlackJack && isParticipantBlackJack) {
+			return Versus.WIN;
+		}
+		return Versus.LOSE;
+	}
+
+	public Versus finalCompare(Participant other) {
+		if (isBust()) {
+			return Versus.LOSE;
+		}
+		if (other.isBust()) {
+			return Versus.WIN;
+		}
+		return whoIsHigh(other.getBestScore());
+	}
+
+	private Versus whoIsHigh(int otherScore) {
+		int myScore = getBestScore();
+		if (myScore > otherScore) {
+			return Versus.WIN;
+		}
+		if (myScore == otherScore) {
+			return Versus.DRAW;
+		}
+		return Versus.LOSE;
+	}
+
+	public boolean compareName(String name) {
+		return this.name.equals(name);
+	}
+}
